requirements.txt
python-dotenv
termcolor

langchain
langchain_openai
langchain_community

openai
anthropic

gradio
faiss-cpu

flask
fastapi
uvicorn

slither-analyzer
echidna